+++ common/src/main/java/net/spell_engine/SpellEngineMod.java
@@ -22,7 +22,7 @@ import net.spell_engine.entity.SpellProjectile;
 import net.spell_engine.internals.SpellContainerHelper;
 import net.spell_engine.internals.SpellRegistry;
 import net.spell_engine.internals.criteria.EnchantmentSpecificCriteria;
-import net.spell_engine.internals.criteria.SpellCastCriteria;
+import net.spell_engine.spellbinding.SpellBindingCriteria;
 import net.spell_engine.network.ServerNetwork;
 import net.spell_engine.particle.Particles;
 import net.spell_engine.spellbinding.*;
@@ -54,7 +54,6 @@ public class SpellEngineMod {
         ServerNetwork.initializeHandlers();
         Particles.register();
 
-        Criteria.register(SpellCastCriteria.INSTANCE);
         Criteria.register(EnchantmentSpecificCriteria.INSTANCE);
         EnchantmentRestriction.alleviate(Enchantments.KNOCKBACK, itemStack -> itemStack.getItem() instanceof StaffItem);
         EnchantmentRestriction.alleviate(Enchantments.LOOTING, itemStack -> itemStack.getItem() instanceof StaffItem);