+++ common/src/main/java/net/spell_engine/internals/SpellHelper.java
@@ -20,8 +20,6 @@ import net.spell_engine.api.enchantment.Enchantments_SpellEngine;
 import net.spell_engine.api.spell.Spell;
 import net.spell_engine.entity.ConfigurableKnockback;
 import net.spell_engine.entity.SpellProjectile;
-import net.spell_engine.internals.criteria.SpellCastCriteria;
-import net.spell_engine.internals.criteria.SpellCastHistory;
 import net.spell_engine.particle.ParticleHelper;
 import net.spell_engine.utils.AnimationHelper;
 import net.spell_engine.utils.SoundHelper;
@@ -263,10 +261,6 @@ public class SpellHelper {
                 if (spell.cost.effect_id != null) {
                     var effect = Registry.STATUS_EFFECT.get(new Identifier(spell.cost.effect_id));
                     player.removeStatusEffect(effect);
-                }
-                // Advancement support
-                if (player instanceof ServerPlayerEntity serverPlayer) {
-                    ((SpellCastHistory)serverPlayer).saveSpellCast(spellId);
                 }
             }
         }