+++ common/src/main/java/net/spell_engine/internals/criteria/SpellCastHistory.java
@@ -1,9 +1,0 @@
-package net.spell_engine.internals.criteria;
-
-import net.minecraft.util.Identifier;
-
-import java.util.List;
-
-public interface SpellCastHistory {
-    void saveSpellCast(Identifier spell);
-}