+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # 0.9.16
 
+- Add
+- Remove advancements spell cast criteria (due to theoretical poor performance)
 - Set `generic.attack_damage` vanilla attribute to be synchronized over to clients
 - Fix use spell_cost_item_allowed still requiring at least 1 rune
-- Advancements spell cast criteria is now pool based instead of magic school
+
 
 API breaking changes:
 - `spell.impact[].action.status_effect.apply_to_caster` was moved to `spell.impact[].action.apply_to_caster`